requirements.txt
agent-framework==1.0.0b260212
agent-framework-neo4j>=0.5.4
neo4j-agent-memory[microsoft-agent]>=0.0.4
neo4j>=5.20.0
neo4j-graphrag[openai]>=1.13.0
openai>=1.99.0
opentelemetry-semantic-conventions-ai==0.4.13
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
